// src/mail/mails/mails-dto/mail.dto.ts
export class MailDTO {
  title: string;
  orgin: string;
  type: string;
  notes: string;
  fromMail: string;
  inNumber: string;
  outNumber: string;
  inDate: Date;
  outDate: Date;
  date_presenting: Date;
  status: string;
  urgent: boolean;
  urgent_by: string;
  saved_place: string;
  file: string;
  catCatID;
}